% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dictionary.R
\name{apply_orth_rules}
\alias{apply_orth_rules}
\title{Apply orthographic rules to text}
\usage{
apply_orth_rules(text, rules = orth_rules)
}
\arguments{
\item{text}{Character vector with input text.}

\item{rules}{Data frame of orthographic rules. Defaults to
\code{orth_rules}.}
}
\value{
Character vector with replacements applied.
}
\description{
Replaces variant-specific spellings using the built-in \code{orth_rules}
data set. Each rule should be formatted as "from -> to".
}
\examples{
apply_orth_rules("acto", orth_rules)
}
