% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stopwords.R
\name{get_stopwords}
\alias{get_stopwords}
\title{Get Portuguese stopwords}
\usage{
get_stopwords(extra = NULL, omit = NULL)
}
\arguments{
\item{extra}{Optional character vector of custom stopwords to add.}

\item{omit}{Optional character vector of stopwords to remove.}
}
\value{
Character vector of stopwords.
}
\description{
Returns a vector of built-in Portuguese stopwords with optional
additions or removals.
}
\examples{
get_stopwords()
get_stopwords(extra = c("novapalavra"))
}
