% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/social.R
\name{normalize_accents}
\alias{normalize_accents}
\title{Normalize accents in Portuguese text}
\usage{
normalize_accents(text)
}
\arguments{
\item{text}{Character vector with input text.}
}
\value{
Character vector with accents normalized.
}
\description{
Converts accented characters to their ASCII equivalents using
\code{stringi::stri_trans_general}.
}
\examples{
normalize_accents("ação")
}
