% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.R
\name{normalize_text}
\alias{normalize_text}
\title{Normalize Portuguese text}
\usage{
normalize_text(text, lower = TRUE, remove_punct = FALSE, correct = FALSE)
}
\arguments{
\item{text}{Character vector with input text.}

\item{lower}{Logical indicating whether to convert to lower case. Defaults to
\code{TRUE}.}

\item{remove_punct}{Logical indicating whether to remove punctuation. Defaults
to \code{FALSE}.}

\item{correct}{Logical indicating whether to apply built-in orthographic
rules before accent normalization.}
}
\value{
A character vector with normalized text.
}
\description{
This function applies basic normalization steps to Portuguese text, such as
converting to lower case, removing punctuation, and folding accents using
\code{stringi} utilities.
}
\examples{
normalize_text("acção", correct = TRUE)

}
