% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stopwords.R
\docType{data}
\name{stopwords_pt}
\alias{stopwords_pt}
\title{Portuguese stopword list}
\format{
Character vector of stopwords.
}
\source{
Combined from the NLTK list and \url{https://github.com/stopwords-iso/stopwords-pt}.
}
\usage{
stopwords_pt
}
\description{
A small sample of common Portuguese stopwords used by \code{get_stopwords()}.
}
\examples{
data(stopwords_pt)
}
\keyword{datasets}
